HV: cfg: introduce maximum of VMs, vCPUs, etc. as configurations
Malloc() is quite widely used for allocating descriptors for VMs, vCPUs, etc. To
remove malloc(), we need to define a global array for each kind of such
descriptor, and such arrays must have a fixed size.

This implies that the maximum number of supported VMs or vCPUs will additionally
depend on the configurations, which may be unacceptable for general-purpose
builds. Thus a macro CONFIG_NO_MALLOC is introduced to control whether this
build should be heap-free or not.

Notes:

1. The actual configuration values in this patch are solely for illustration.
Shall be adjusted according to the requirements.

2. It is still under discussion how we should improve our building system and
document these configuration entries. Detailed descriptions will be added after
the choosed configuration tool is setup.

Change-Id: I2db2572a2e5eadde5ff829a42ef7472a5d9265d4
Signed-off-by: Junjie Mao <junjie.mao@intel.com>