HV: startup: panic on exceeding the configured maximum of cores
The heap-free hypervisor has to panic if there are more physical cores than the
configured maximum (since the APs are brought up by broadcast INIT/START IPIs
currently). This is an instance of the impact of the introduced hard limit to
the number of supported CPUs after heap is removed.

A possible solution is to bring up APs one after another, at the cost of booting
time which is a critical KPI for the hypervisor.

Note:

1. This patch is for illustration only. The added panic() will not print
anything because logmsg is not yet initialized when initializing per_cpu data
sections.

2. Currently CONFIG_NR_CPUS is valid only when CONFIG_NO_MALLOC is set. The
ifdef around the check can be removed if CONFIG_NR_CPUS is made generic for all
platforms. Whether the other CONFIG_NR_XXX configurations should be generic or
heap-free specific should be analyzed case by case.

v1 -> v2:

    * Add a note to the commit log on the generality of CONFIG_NR_XXX macros.

Change-Id: Ie737a9cf73ddff6c77ebfb4bb8534e2b4ea2ce89
Signed-off-by: Junjie Mao <junjie.mao@intel.com>