HV: irq: replace dynamically allocated irq desc table with a global array
The irq descriptor table is currently dynamically allocated.

Since there is already an upper limit to the number of IRQs we support, it is
straight-forward to convert irq_desc_base to a global array. This is done
regardless whether CONFIG_NO_MALLOC is defined because no additional restriction
is introduced.

Note:

1. The alignment is derived from the current allocation function but it can to
over restrictive.

Change-Id: I5696b5b7398ddb36afba3b9d3954aa597299f106
Signed-off-by: Junjie Mao <junjie.mao@intel.com>