HV: irq: replace dynamically allocated nodes with a dedicated pool
The dev_handler_nodes (linked in each irq_desc) are currently dynamically
allocated. This patch introduces a dedicated, bitmap-based dev_handler_node pool
for each irq_desc to remove the calloc() call. The capacity of the pools are
configured at compile time.

The following two functions are introduced to wrap the underlying difference in
allocation.

  - alloc_dev_handler_node(int irq)
  - free_dev_handler_node(int irq, struct dev_handler_node *node)

When CONFIG_NO_MALLOC is unset, they simply maps to malloc/free calls, while
with CONFIG_NO_MALLOC set they operates on the bitmap of the dedicated pool.

Note:

1. A side effect of switching to a pool-based solution is that dev_handler_nodes
are no longer guaranteed to be 16-byte aligned. Explicit padding is needed if
the alignment is performance-critical.

2. The bitmap is an uint64_t instead of an uint64_t array for simplicity of this
illustration patch. A more general solution (with MISRA C restrictions in mind)
is WIP.

v1 -> v2:

    * Wrap the allocation with a new API alloc_per_cpu_data() which calls
      malloc() when heap is available, while pick from a dedicated pool
      when configured to be heap-free.

    * With CONFIG_NO_MALLOC, modifications to the same pool is synchronized by
      atomic operations instead of locks.

Change-Id: I9c519c1471029796e5771ed568fe61539a423eb5
Signed-off-by: Junjie Mao <junjie.mao@intel.com>