slice: scenario: REVERT ME: restrict the linux VM to 1 core
For now a linux VM with a single vCPU runs normally, while assigning multiple
vCPUs to it triggers random hangs upon AP initialization or kernel
OOPS. Temporarily reduce the number of vCPUs assigned to 1 for the linux VM for
test purposes.

This patch shall be reverted once the kernel hang and OOPS issues are resolved.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>