config_tools: add a slicer of XML schema
We plan to add the following attributes to element definitions in the XML
schema:

  - acrn:applicable-vms, which specify if an element applies to a pre-launched
    VM, the service VM or a post-launched VM

  - acrn:views, which specify if an element shall appear in the basic or
    advanced tab in the configurator.

In order to reduce the attributes above to existing XML technologies, we need to
create new complex types that lists all config items that applies to a
pre-launched VM, the service VM or a post-launched VM, or that should be shown
in the basic or advanced view. Such types can then be used to replace the
original, all-in-one type during validation or configurator rendering.

To accompolish this, this patch adds a generic class implementing the common
part of slicing XML schema types and two specific-purpose slicers according to
the applicable VMs or views attributes.

Tracked-On: #6690
Signed-off-by: Junjie Mao <junjie.mao@intel.com>