config_tools: allow specifying the source of options
This patch refines the syntax of acrn:options annotation by prepending a
prefix to the original XPATH, separated by a colon (':'). The prefix is
used to specify the source on which the following path shall be evaluated.

With this prefix, it is now possible to not only specify an XPATH to be
evaluated on a board XML, but also a JSON schema on a jsonized scenario
data object. This mechanism will thus be reused to specify the options
where VM names are needed.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>