config_tools: combine requirements.txt for config_tools
Today we have two requirement.txt files, one under misc/config_tools/ and
the other misc/config_tools/configurator/. The former was introduced to
capture the dependencies of the Python scripts invoked at build time and
the latter of the configurator.

This patch combines those two files as the scripts and the configurator are
both parts of the ACRN configuration toolset. There is no need to maintain
separate dependencies for each of them.

The requirements.txt under doc/ is not touched as it has additional
dependencies as well as requiring specific versions of them for
compatibility.

Tracked-On: #6690
Signed-off-by: Junjie Mao <junjie.mao@intel.com>